+++ backend/canisters/group/api/src/updates/c2c_join_group.rs
@@ -6,6 +6,7 @@ use types::{GateCheckFailedReason, GroupCanisterGroupChatSummary, UserId};
 pub struct Args {
     pub user_id: UserId,
     pub principal: Principal,
+    pub invite_code: Option<u64>,
     pub correlation_id: u64,
     pub is_platform_moderator: bool,
 }