+++ backend/canisters/group/CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [unreleased]
 
+### Added
+
+- Re-introduce invite by code on backend ([#3552](https://github.com/open-ic/open-chat/pull/3552))
+
 ### Changed
 
 - Short circuit without calling `ic0.time()` if there have been no updates ([#3539](https://github.com/open-ic/open-chat/pull/3539))